Remove flag name check & add boloStatusUid
On BOLO checking it checks the flags section, it checks it's category ID (1) and the label of the section, since the label of the section can be changed per community for translations, personal preference, etc. so it shouldn't be restricted to just "Flags" as the label.

Additionally adds config option to specify the bolo status uid as that can be different than just "status". This is in the lookup plugin instead of where it's used (wraithv2 plugin) as here it's evaluated if it's a valid BOLO or not.